Finished
1) A setCoordinates() function was added, setting coordinate1 and
coordinate2 based on the played move. These coordinates represent spaces
on self.board.
2) Changed checkValidMove() to isValidMove(). isValidMove() now checked
the coordinates instead of the actual move.
3) A new function, placeMove() was added which plays the move on
self.board determined by coordinate1 and coordinate2.
4) Added startGame() function. This function starts the game, draws the
board, and asks the player to input a move. It checks for a valid move
and redraws the board until a player has won.
5) Removed test cases.